introduce CheckCycle to detect cycles in project's depends_on
